AI Workflow Functions/fetch_stock_data/requirements.txt
functions-framework
yfinance
pandas
cloud-sql-python-connector[pg8000]
sqlalchemy
pg8000
requests
google-cloud-secret-manager